use float64 dtype for array

+++ omics_graph_learning/graph_constructor.py
@@ -161,8 +161,10 @@ def _nx_to_tensors(
     graph = nx.convert_node_labels_to_integers(graph, ordering="sorted")
     edges = np.array([[edge[0], edge[1]] for edge in nx.to_edgelist(graph)]).T
     node_features = np.asarray(
-        [list(graph.nodes[node].values()) for node in sorted(graph.nodes)]
+        [list(graph.nodes[node].values()) for node in sorted(graph.nodes)],
+        dtype="float64",
     )
+
     edge_features = [graph[u][v][2] for u, v in edges.T]
 
     with open(f"{graph_dir}/{prefix}_{graph_type}_graph_{tissue}.pkl", "wb") as output:
